+++ formatters/string.js
@@ -312,6 +312,24 @@ function append (d, append) {
 
 
 /**
+ * Neutral for array filters
+ *
+ * This formatter can be used to "ignore" Carbone tags if an array filter is used.
+ * For example:
+ *   XML: {d[i, id>0].id} {c.now:neutralForArrayFilter:somethingElse} {d[i, id>0].name}
+ *
+ * Array filters works only if all Carbone tags are filtered. But `{c.now}` has no filter, so it will print all
+ * rows of the d[] array because there is at least one Carbone tag without a filter.
+ * With :neutralForArrayFilter formatter, we can tell to Carbone: "filter array rows regardless the presence of this Carbone tag"
+ *
+ * @private
+ */
+function neutralForArrayFilter (d) {
+  return d;
+}
+
+
+/**
  * Returns the length of a string or array.
  *
  * @version 2.0.0
@@ -347,6 +365,7 @@ module.exports = {
   md5       : md5,
   prepend   : prepend,
   append    : append,
+  neutralForArrayFilter : neutralForArrayFilter,
   // private
   convCRLFH : convCRLFH
 };